+++ src/main/java/com/soutenence/publiciteApp/specification/ImageSpecification.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.domain.Specification;
 
 public class ImageSpecification {
 
-    public static Specification<Image> byAbonnement(Integer abonnementId){
+    public static Specification<Image> byAbonnement(Long abonnementId){
         return ((root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get("abonnement").get("id"),abonnementId));
     }
 }
